Introduce EvaluationContext for providers
EvaluationContext provides means to track providers that are being
currently evaluated by the current thread.

With this class, we intend to use try-with-resources to manage the
scopes; however, only opening and closing the scope are important, so
these are going to be mostly unused inside the try block. To make the
life simpler, the Javac warning that complains about unused resources is
disabled. FWIW, Java 22 will finally allow unnamed resources in the try
block, so this is idiomatic.